- Refactored all battle scenes to support more than 1 player - Battle scenes also integrate form tracking data for each player - Many states rewritten + cleaned up to support these new changes - Forms will NOT crash now - Fixed bug in animation component skipping  calls when stunned - Tweaked how I was tracking decrossing so navis who **are not** transforming will not flash white - Added MANY missing element weaknesses to the super-effective table - REMOVED  element as it is not a BN element ever - negative damage will not show up in the text at the bottom-left of the screen - Player spawning is now SYNCHRONIZED for PVP! This means you will always see the same screen even if you have custom NCP parts that modify the field! - Many other crashes and bugs spotted and fixed

+++ BattleNetwork/battlescene/States/bnCardComboBattleState.cpp
@@ -53,7 +53,7 @@ void CardComboBattleState::Simulate(double elapsed, std::vector<Battle::Card>& c
       listStepCounter -= (float)elapsed;
     }
     else {
-      int cardsLen = cards.size();
+      int cardsLen = (int)cards.size();
 
       // +2 = 1 step for showing PA label and 1 step for showing merged card
       // That's the cards we want to show + 1 + 1 = cardCount + 2
